Refactor navbar and update account, home, and messages tabs
