+++ rideshare-app/app/(tabs)/messages/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, StyleSheet, SafeAreaView, Platform } from 'react-native';
 import { colors } from '../../../ui/styles/colors';
-import NavBar from '../../../src/components/nav-bar';
+import NavBar from '../../../app/components/nav-bar';
 
 export default function MessagesScreen() {
   return (
