+++ rideshare-app/app/components/nav-bar.js
